+++ pkg/config/config.go
@@ -8,16 +8,18 @@ import (
 // K3sVersion default version
 const K3sVersion = "v0.8.1"
 
-func InitUserDir() error {
+func InitUserDir() (string, error) {
 	home := os.Getenv("HOME")
+	fullPath := fmt.Sprintf("%s/k3sup/.bin/", home)
+
 	if len(home) == 0 {
-		return fmt.Errorf("env-var HOME, not set")
+		return fullPath, fmt.Errorf("env-var HOME, not set")
 	}
 
-	err := os.MkdirAll(fmt.Sprintf("%s/k3sup/.bin/", home), 0700)
+	err := os.MkdirAll(fullPath, 0700)
 	if err != nil {
-		return err
+		return fullPath, err
 	}
 
-	return nil
+	return fullPath, nil
 }
